+++ clients/admin-ui/__tests__/jest.setup.ts
@@ -1,2 +1,5 @@
 import "@testing-library/jest-dom";
 import "whatwg-fetch";
+
+// eslint-disable-next-line global-require
+jest.mock("iso-3166", () => require("./utils/iso-3166-mock").iso3166Mock);